update CRUD for tool in admin dashboard

+++ src/main/java/com/example/Project/repository/ICategoryRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.Project.model.Category;
 
-public interface ICategoryRepository extends JpaRepository<Category, Long> {
+public interface ICategoryRepository extends JpaRepository<Category, Integer> {
 }
